requirements.txt
django
djangorestframework
pytest-django
pyyaml
requests